requirements.txt
# Core
# -------------------------------------
django-oscar==1.3
django-oscar-paypal==0.9.7
Django==1.9.9
gunicorn==19.6.0

# Payments
# -------------------------------------
dj-stripe==0.8.0

# Audio Conversion
# -------------------------------------
pydub==0.16.5

# Extensions
# -------------------------------------
dj-database-url==0.4.1
django-extensions==1.7.1
django-uuid-upload-path==1.0.0
pytz==2016.6.1
ipython==5.1.0
pycountry==1.20
Werkzeug==0.10.4

# Configuration
# -------------------------------------
django-environ==0.4.0
python-dotenv==0.5.1

# Images
# -------------------------------------
Pillow==3.0.0

# Staticfiles
# -------------------------------------
dj-static==0.0.6
django-compressor==2.1

# Persistence
# -------------------------------------
dj-database-url==0.4.1
psycopg2==2.6.2

# Process
# -------------------------------------
honcho==0.7.1
